contributors/ttwf_shenzhen/css-shapes/sherlock/submitted/shape-outside-invalid-circle-003.html

<!doctype html>
<html>
<head>
	<meta charset="UTF-8">
	<meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if circle shape function has an invalid cy param">
	<title>CSS Shapes Test: Circle - invalid cx value</title>
	<link rel="author" title="Sherlock Zhong" href="mailto:sherlock.z0917@gmail.com">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<link rel="stylesheet" href="/resources/testharness.css">
	<style>
#shape{
	width: 300px;
	height: 300px;
	shape-outside: circle(50, 50%, 50%);
}
	</style>
</head>
<body>

	<div id="log"></div>

	<div id="shape"></div>

	<script>
	var obj = document.getElementById('shape');
	var cx = getComputedStyle(obj).shapeOutside;
	test(function() {assert_equals(cx, "auto", "declaration should be invalid")});
	</script>

</body>
</html>
